Change user controller to admin path
